Feature/star and long in entities

+++ src/Company.Videomatic.Application/Features/Artifacts/ArtifactDTO.cs
@@ -1,8 +1,8 @@
 ﻿namespace Company.Videomatic.Application.Features.Artifacts;
 
 public record ArtifactDTO( 
-    long ArtifactId = 0,
-    long VideoId = 0,
+    int ArtifactId = 0,
+    int VideoId = 0,
     string Name = "",
     string Type = "",
     string? Text = null);
