+++ src/Company.Videomatic.Application/Features/Playlists/PlaylistDTO.cs
@@ -1,8 +1,8 @@
 ﻿namespace Company.Videomatic.Application.Features.Playlists;
 
 public record PlaylistDTO(
-    long Id = 0,
+    int Id = 0,
     string Name = "",
     string? Description = default,
 
-    long? VideoCount = 0);
+    int? VideoCount = 0);
