+++ src/Company.Videomatic.Application/Features/Videos/VideoDTO.cs
@@ -1,7 +1,7 @@
 ﻿namespace Company.Videomatic.Application.Features.Videos;
 
 public record VideoDTO(
-    long Id = 0,
+    int Id = 0,
     string Location = "",
     string Name = "",
     string? Description = null,
