+++ src/Company.Videomatic.Application/Features/Videos/VideoIdAndProviderVideoIdDTO.cs
@@ -1,3 +1,3 @@
 ﻿namespace Company.Videomatic.Application.Features.Videos.Queries;
 
-public record VideoIdAndProviderVideoIdDTO(long VideoId, string ProviderVideoId);
+public record VideoIdAndProviderVideoIdDTO(int VideoId, string ProviderVideoId);
