+++ src/Company.Videomatic.Domain/Aggregates/Artifact/ArtifactId.cs
@@ -1,7 +1,7 @@
 ﻿namespace Company.Videomatic.Domain.Aggregates.Artifact;
 
-public record ArtifactId(long Value = 0)
+public record ArtifactId(int Value = 0)
 {
-    public static implicit operator long(ArtifactId x) => x.Value;
-    public static implicit operator ArtifactId(long x) => new (x);
+    public static implicit operator int(ArtifactId x) => x.Value;
+    public static implicit operator ArtifactId(int x) => new (x);
 }
