+++ tests/Application.Tests/Validation/ArtifactValidatorTests.cs
@@ -20,7 +20,7 @@ public class ArtifactValidatorTests
     [InlineData(3, null, "Name", "Text", 1)]
     [InlineData(4, "AI", null, "Text", 1)]
     [InlineData(5, null, null, null, 3)]
-    public void ValidateCreateArtifactCommand(long videoId, string type, string name, string? text, int expectedErrors)
+    public void ValidateCreateArtifactCommand(int videoId, string type, string name, string? text, int expectedErrors)
     {
         ValidatorHelper.Validate<CreateArtifactCommandValidator, CreateArtifactCommand>(new(videoId, name, type, text), expectedErrors);
     }
@@ -29,7 +29,7 @@ public class ArtifactValidatorTests
     [InlineData(0, 1)]
     [InlineData(-1, 1)]
     [InlineData(1, 0)]
-    public void ValidateDeletePlaylistCommand(long id, int expectedErrors)
+    public void ValidateDeletePlaylistCommand(int id, int expectedErrors)
     {
         ValidatorHelper.Validate<DeletePlaylistCommandValidator, DeletePlaylistCommand>(new(id), expectedErrors);
     }
@@ -40,7 +40,7 @@ public class ArtifactValidatorTests
     [InlineData(1, null, null, 1)]
     [InlineData(1, "Play list", null, 0)]
     [InlineData(2, "Play list", "Description", 0)]
-    public void ValidateUpdatePlaylistCommand(long id, string name, string? description, int expectedErrors)
+    public void ValidateUpdatePlaylistCommand(int id, string name, string? description, int expectedErrors)
     {
         ValidatorHelper.Validate<UpdatePlaylistCommandValidator, UpdatePlaylistCommand>(new(id, name, description), expectedErrors);
     }
